agogosml/requirements.txt
confluent-kafka~=0.11
azure-eventhub~=1.2
requests~=2.20
redis~=3.1.0
python-dotenv~=0.9
flask~=1.0.2
pyyaml>=4.2b1
jsonschema~=2.6
applicationinsights~=0.11.7
cached-property~=1.5.1
singleton-decorator~=1.0.0
apache-libcloud~=2.4.0
